# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted mail platform — REST API for the admin console and webmail.",
        "title": "mailez API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/ai/draft": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Draft reply",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
        "/ai/prioritize": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Prioritize inbox",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
        "/ai/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "AI search",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mailez_backend_internal_mail.Message"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/ai/summarize": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Summarize email",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/alternatives": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "List alternatives",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mailez_backend_internal_core_models.Alternative"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "Create alternative",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.Alternative"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
        "/alternatives/{name}": {
            "delete": {
                "tags": [
                    "domains"
                ],
                "summary": "Delete alternative",
                "parameters": [
                    {
                        "type": "string",
                        "description": "alternative name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/domains/{name}/dkim": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "DKIM status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "Generate DKIM key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mailez_backend_internal_core_models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "rate_limited"
                },
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "mailez_backend_internal_core_models.Alternative": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "domain_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "mailez_backend_internal_mail.Address": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "mailez_backend_internal_mail.Attachment": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "mailez_backend_internal_mail.Message": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailez_backend_internal_mail.Attachment"
                    }
                },
                "cc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailez_backend_internal_mail.Address"
                    }
                },
                "date": {
                    "type": "string"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "folder": {
                    "type": "string"
                },
                "from": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailez_backend_internal_mail.Address"
                    }
                },
                "has_attachment": {
                    "type": "boolean"
                },
                "html_body": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "seq": {
                    "type": "integer"
                },
                "subject": {
                    "type": "string"
                },
                "text_body": {
                    "type": "string"
                },
                "thread_count": {
                    "type": "integer"
                },
                "thread_id": {
                    "type": "string"
                },
                "thread_latest": {
                    "type": "boolean"
                },
                "to": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailez_backend_internal_mail.Address"
                    }
                },
                "uid": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "mailez_session",
            "in": "cookie"
        }
    }
}